use enums in request headers
Promotes the magicvalue and opcode enums higher up in the binary
module. Refactors request header and parsing logic to use these
types.
