+++ src/protocol/memcache/src/response/server_error.rs
@@ -2,7 +2,8 @@
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
-use crate::binary::response::header::{Opcode, ResponseStatus};
+use crate::binary::Opcode;
+use crate::binary::response::header::{ResponseStatus};
 
 use super::*;
